requirements.txt
-i https://pypi.douban.com/simple
astartool
numpy~=1.22.3
matplotlib~=3.5.1
Pillow~=9.1.0
opencv-python
numpy
pandas
pyproj
requests
selenium
password
prettytable
snowland-algorithm
bitarray
pygame
cfg2
modules
#pip3 install -r requirements.txt -i https://pypi.douban.com/simple